[fs] sunrpc: if we're closing down a socket, clear memalloc on it first
Message-id: <c79b4e9cdf15163a2447db6ae6db5348bb9c9383.1457439311.git.bcodding@redhat.com>
Patchwork-id: 137450
O-Subject: [RHEL7 PATCH fs 141/396] sunrpc: if we're closing down a socket, clear memalloc on it first
Bugzilla: 1315390
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Jeff Layton <jlayton@poochiereds.net>

We currently increment the memalloc_socks counter if we have a xprt that
is associated with a swapfile. That socket can be replaced however
during a reconnect event, and the memalloc_socks counter is never
decremented if that occurs.

When tearing down a xprt socket, check to see if the xprt is set up for
swapping and sk_clear_memalloc before releasing the socket if so.

Acked-by: Mel Gorman <mgorman@suse.de>
Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Reviewed-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>

Upstream-Id: 264d1df3b34804a7d440d77771020f616a573528
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1315390
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>