[fs] nfs: make nfs4_init_nonuniform_client_string use a dynamically allocated buffer
Message-id: <e3b99fc6cd76df0f6ae55f004f18fd59ee8a953b.1457439311.git.bcodding@redhat.com>
Patchwork-id: 137299
O-Subject: [RHEL7 PATCH fs 169/396] nfs: make nfs4_init_nonuniform_client_string use a dynamically allocated buffer
Bugzilla: 1315390
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Jeff Layton <jlayton@poochiereds.net>

The way the *_client_string functions work is a little goofy. They build
the string in an on-stack buffer and then use kstrdup to copy it. This
is not only stack-heavy but artificially limits the size of the client
name string. Change it so that we determine the length of the string,
allocate it and then scnprintf into it.

Since the contents of the nonuniform string depend on rcu-managed data
structures, it's possible that they'll change between when we allocate
the string and when we go to fill it. If that happens, free the string,
recalculate the length and try again. If it the mismatch isn't resolved
on the second try then just give up and return -EINVAL.

Signed-off-by: Jeff Layton <jeff.layton@primarydata.com>
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream-Id: a319268891089386bd303bb90ace89d75e47102f
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1315390
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>