[fs] nfs: Don't revalidate the mapping if both size and change attr are up to date
Message-id: <a0513041fa63301ea1cca9579d999df6872d5b2e.1457439311.git.bcodding@redhat.com>
Patchwork-id: 137518
O-Subject: [RHEL7 PATCH fs 223/396] NFS: Don't revalidate the mapping if both size and change attr are up to date
Bugzilla: 1315390
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

If we've ensured that the size and the change attribute are both correct,
then there is no point in marking those attributes as needing revalidation
again. Only do so if we know the size is incorrect and was not updated.

Fixes: f2467b6f64da ("NFS: Clear NFS_INO_REVAL_PAGECACHE when...")
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream-Id: 85a23cee3f2c928475f31777ead5a71340a12fc3
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1315390
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>