[fs] nfsv4: Recovery of recalled read delegations is broken
Message-id: <798f3cebb652564f2fe11443a317880c9061c38c.1457439311.git.bcodding@redhat.com>
Patchwork-id: 137559
O-Subject: [RHEL7 PATCH fs 342/396] NFSv4: Recovery of recalled read delegations is broken
Bugzilla: 1315390
RH-Acked-by: Steve Dickson <SteveD@redhat.com>
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

From: Trond Myklebust <trond.myklebust@primarydata.com>

When a read delegation is being recalled, and we're reclaiming the
cached opens, we need to make sure that we only reclaim read-only
modes.
A previous attempt to do this, relied on retrieving the delegation
type from the nfs4_opendata structure. Unfortunately, as Kinglong
pointed out, this field can only be set when performing reboot recovery.

Furthermore, if we call nfs4_open_recover(), then we end up clobbering
the state->flags for all modes that we're not recovering...

The fix is to have the delegation recall code pass this information
to the recovery call, and then refactor the recovery code so that
nfs4_open_delegation_recall() does not need to call nfs4_open_recover().

Reported-by: Kinglong Mee <kinglongmee@gmail.com>
Fixes: 39f897fdbd46 ("NFSv4: When returning a delegation, don't...")
Tested-by: Kinglong Mee <kinglongmee@gmail.com>
Cc: NeilBrown <neilb@suse.com>
Cc: stable@vger.kernel.org # v4.2+
Signed-off-by: Trond Myklebust <trond.myklebust@primarydata.com>
Upstream-Id: 24311f884189d42d40354a6f38ca218eb9aeb811
Bugzilla: https://bugzilla.redhat.com/show_bug.cgi?id=1315390
Signed-off-by: Benjamin Coddington <bcodding@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>