[include] perf/core: Drop PERF_EVENT_TXN
Message-id: <1457464300-22518-23-git-send-email-jolsa@redhat.com>
Patchwork-id: 137598
O-Subject: [PATCH RHEL7.3 BZ1274454 22/22] perf/core: Drop PERF_EVENT_TXN
Bugzilla: 1274454
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

From: Jiri Olsa <jolsa@kernel.org>

Bugzilla: 1274454
https://bugzilla.redhat.com/show_bug.cgi?id=1274454

upstream
========
commit 8f3e5684d3fbd91ead283916676fa3dac22615e5
Author: Sukadev Bhattiprolu <sukadev@linux.vnet.ibm.com>
Date: Thu Sep 3 20:07:53 2015 -0700

RHEL7
=====
Ommited the sparc code hunk.

description
===========
We currently use PERF_EVENT_TXN flag to determine if we are in the middle
of a transaction. If in a transaction, we defer the schedulability checks
from pmu->add() operation to the pmu->commit() operation.

Now that we have "transaction types" (PERF_PMU_TXN_ADD, PERF_PMU_TXN_READ)
we can use the type to determine if we are in a transaction and drop the
PERF_EVENT_TXN flag.

When PERF_EVENT_TXN is dropped, the cpuhw->group_flag on some architectures
becomes unused, so drop that field as well.

This is an extension of the Powerpc patch from Peter Zijlstra to s390,
Sparc and x86 architectures.

Signed-off-by: Rafael Aquini <aquini@redhat.com>