[mm] soft-dirty bits for user memory changes tracking
Message-id: <20160308205334.GA17478@redhat.com>
Patchwork-id: 137959
O-Subject: [RHEL7 PATCH BZ 1269561 05/21] mm: soft-dirty bits for user memory changes tracking
Bugzilla: 1269561
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Adrian Reber <areber@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

RHBZ: 1269561

The soft-dirty is a bit on a PTE which helps to track which pages a task
writes to.  In order to do this tracking one should

  1. Clear soft-dirty bits from PTEs ("echo 4 > /proc/PID/clear_refs)
  2. Wait some time.
  3. Read soft-dirty bits (55'th in /proc/PID/pagemap2 entries)

To do this tracking, the writable bit is cleared from PTEs when the
soft-dirty bit is.  Thus, after this, when the task tries to modify a
page at some virtual address the #PF occurs and the kernel sets the
soft-dirty bit on the respective PTE.

Note, that although all the task's address space is marked as r/o after
the soft-dirty bits clear, the #PF-s that occur after that are processed
fast.  This is so, since the pages are still mapped to physical memory,
and thus all the kernel does is finds this fact out and puts back
writable, dirty and soft-dirty bits on the PTE.

Another thing to note, is that when mremap moves PTEs they are marked
with soft-dirty as well, since from the user perspective mremap modifies
the virtual memory at mremap's new address.

FIXUPS:
	conflict in pagemap_pte_range(), pmd_trans_huge_lock() has
	additional "ptl" arg compared to upstream

	conflict in move_huge_pmd() caused by f1025d75 (f1025d75d17da
	upstream)

	misc trivial conflicts in */Kconfig

Signed-off-by: Pavel Emelyanov <xemul@parallels.com>
Cc: Matt Mackall <mpm@selenic.com>
Cc: Xiao Guangrong <xiaoguangrong@linux.vnet.ibm.com>
Cc: Glauber Costa <glommer@parallels.com>
Cc: Marcelo Tosatti <mtosatti@redhat.com>
Cc: KOSAKI Motohiro <kosaki.motohiro@gmail.com>
Cc: Stephen Rothwell <sfr@canb.auug.org.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit 0f8975ec4db2c8b5bd111b211292ca9be0feb6b8)
Signed-off-by: Rafael Aquini <aquini@redhat.com>