[crypto] crc32 - Rename generic implementation
Message-id: <20160309135702.GA3271@gondor.apana.org.au>
Patchwork-id: 138048
O-Subject: [RHEL7 PATCH] crypto: crc32 - Rename generic implementation
Bugzilla: 1314773
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHEL7 Bugzilla #1314773

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=10627817

Upstream commit: a7c58ac06224b83dba93eaafde88535d5da60aef

The generic crc32 implementation is currently called crc32.  This
is a problem because it clashes with the lib implementation of crc32.

This patch renames the crypto crc32 to crc32_generic so that it is
consistent with crc32c.  An alias for the driver is also added.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Rafael Aquini <aquini@redhat.com>