[scsi] only re-lock door after EH on devices that were reset
Message-id: <1457615955-3985-1-git-send-email-emilne@redhat.com>
Patchwork-id: 138279
O-Subject: [RHEL7.3 PATCH e-stor] scsi: only re-lock door after EH on devices that were reset
Bugzilla: 1316290
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>

From upstream linux-stable

Tested at customer site on their TANDBERG RDX scsi device

Note:  Although this commit message refers to the blk-mq I/O path,
this patch fixes a problem where attempting to submit a SCSI command
to re-lock the door hangs in blk_get_request() because the queue is
already full, regardless of whether scsi-mq is enabled.  This solved
the problem for the customer.  The patch is in RHEL6 and must be
included in RHEL7 to avoid a regression in the event of an upgrade.

upstream commit 48379270fe6808cf4612ee094adc8da2b7a83baa
Author: Christoph Hellwig <hch@lst.de>
Date:   Mon Nov 3 19:36:40 2014 +0100

    scsi: only re-lock door after EH on devices that were reset

    Setups that use the blk-mq I/O path can lock up if a host with a single
    device that has its door locked enters EH.  Make sure to only send the
    command to re-lock the door to devices that actually were reset and thus
    might have lost their state.  Otherwise the EH code might be get blocked
    on blk_get_request as all requests for non-reset devices might be in use.

    Cc: stable@vger.kernel.org
    Signed-off-by: Christoph Hellwig <hch@lst.de>
    Reported-by: Meelis Roos <meelis.roos@ut.ee>
    Tested-by: Meelis Roos <meelis.roos@ut.ee>
    Reviewed-by: Martin K. Petersen <martin.petersen@oracle.com>

Signed-off-by: Ewan D. Milne <emilne@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>