[include] perf/x86: Add option to disable reading branch flags/cycles
Message-id: <1457691394-28049-4-git-send-email-jolsa@redhat.com>
Patchwork-id: 138312
O-Subject: [PATCH RHEL7.3 BZ1316684 3/5] perf/x86: Add option to disable reading branch flags/cycles
Bugzilla: 1316684
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: David Arcari <darcari@redhat.com>

Bugzilla: 1316684
https://bugzilla.redhat.com/show_bug.cgi?id=1316684

upstream
========
commit b16a5b52eb90d92b597257778e51e1fdc6423e64
Author: Andi Kleen <ak@linux.intel.com>
Date: Tue Oct 20 11:46:34 2015 -0700

description
===========
With LBRv5 reading the extra LBR flags like mispredict, TSX, cycles is
not free anymore, as it has moved to a separate MSR.

For callstack mode we don't need any of this information; so we can
avoid the unnecessary MSR read. Add flags to the perf interface where
perf record can request not collecting this information.

Add branch_sample_type flags for CYCLES and FLAGS. It's a bit unusual
for branch_sample_types to be negative (disable), not positive (enable),
but since the legacy ABI reported the flags we need some form of
explicit disabling to avoid breaking the ABI.

After we have the flags the x86 perf code can keep track if any users
need the flags. If noone needs it the information is not collected.

This cuts down the cost of LBR callstack on Skylake significantly.
Profiling a kernel build with LBR call stack the average run time of
the PMI handler drops by 43%.

Signed-off-by: Rafael Aquini <aquini@redhat.com>