[watchdog] hpwdt: Add support for WDIOC_SETOPTIONS
Message-id: <1457966105-26161-1-git-send-email-ncroxon@redhat.com>
Patchwork-id: 138406
O-Subject: [RHEL7.3 PATCH 1/1] watchdog: hpwdt: Add support for WDIOC_SETOPTIONS
Bugzilla: 1316536
RH-Acked-by: Steve Best <sbest@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Bugzilla: 1316536
https://bugzilla.redhat.com/show_bug.cgi?id=1316536

Brew: 10630816
https://brewweb.devel.redhat.com/taskinfo?taskID=10630816

Description:
Add support for WDIOC_SETOPTIONS in hpwdt (HP Watchdog Timer).
WDIOC_SETOPTIONS makes it possible to disable and re-enable the
watchdog timer while the hpwdt driver is loaded.
This is needed for HPE's next generation platform.

Tested:
On a kernel without this patch, changine the HP watchdog timer to 10s
would print an error msg, "Inappropriate ioctl for device".
A patched kernel was tested on HP platform. The HP watchdog timer
was configured and a call was made to set the hardware watchdog to 10s
and a error messages was not displayed.
A write-up on how to test is posted as a comment to the bz.

-Nigel Croxon
HPE On-site Engineer at Red Hat.

commit 46c80b20d07e9309677d81e1226dc30a055b63b7
Author: Jean Delvare <jdelvare@suse.de>
Date:   Sun Jun 21 09:32:33 2015 +0200

WDIOC_SETOPTIONS makes it possible to disable and re-enable the
watchdog timer while the hpwdt driver is loaded.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Reviewed-by: Guenter Roeck <linux@roeck-us.net>
Signed-off-by: Wim Van Sebroeck <wim@iguana.be>
(cherry picked from commit 46c80b20d07e9309677d81e1226dc30a055b63b7)
Signed-off-by: Rafael Aquini <aquini@redhat.com>