[crypto] gcm - Fix rfc4543 decryption crash
Message-id: <20160323023520.GA25572@gondor.apana.org.au>
Patchwork-id: 139199
O-Subject: [RHEL7 PATCH] crypto: gcm - Fix rfc4543 decryption crash
Bugzilla: 1298642
RH-Acked-by: Linda Wang <lwang@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

RHEL7 Bugzilla #1298642

http://brewweb.devel.redhat.com/brew/taskinfo?taskID=10708647

Upstream commit: adcbc688fe2f8107b7f564187593293aa9ea3932

Note that this is not a backport of the above upstream commit.
The upstream commit inadvertently fixed a bug that is the crux
of the bugzilla entry.  As the upstream commit is not appropriate
for RHEL7, this patch is an independent fix for just this bug.

This fix has been submitted to the stable kernels too.

Original description:

This bug has already bee fixed upstream since 4.2.  However, it
was fixed during the AEAD conversion so no fix was backported to
the older kernels.

When we do an RFC 4543 decryption, we will end up writing the
ICV beyond the end of the dst buffer.  This should lead to a
crash but for some reason it was never noticed.

This patch fixes it by only writing back the ICV for encryption.

Fixes: d733ac90f9fe ("crypto: gcm - fix rfc4543 to handle async...")
Reported-by: Patrick Meyer <patrick.meyer@vasgard.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>