[fs] eventfd: document lockless access in eventfd_poll
Message-id: <1458823337-14717-3-git-send-email-pbonzini@redhat.com>
Patchwork-id: 139241
O-Subject: [RHEL7.3 PATCH v2 2/2] eventfd: document lockless access in eventfd_poll
Bugzilla: 1293632
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Andrea Arcangeli <aarcange@redhat.com>

Since commit e22553e2a25e ("eventfd: don't take the spinlock in
eventfd_poll", 2015-02-17), eventfd is reading ctx->count outside
ctx->wqh.lock.

However, things aren't as simple as the read barrier in eventfd_poll
would suggest.  In fact, the read barrier, besides lacking a comment,
is not paired in any obvious manner with a full or write barrier, and it
is pointless because it is sitting between a write (deep in poll_wait)
and the read of ctx->count.  The read barrier is acting just as a
compiler barrier, for which we can use READ_ONCE instead.  This is
what the code change in this patch does.

The documentation change is just as important, however.  The question,
posed by Andrea Arcangeli, is then why the thing is safe on architectures
where spin_unlock does not imply a store-load memory barrier.
The answer is that it's safe because writes of ctx->count use the same
lock as poll_wait, and hence an acquire barrier implicit in poll_wait
provides the necessary synchronization between eventfd_poll and callers
of wake_up_locked_poll.  This is sort of mentioned in the commit message
with respect to eventfd_ctx_read ("eventfd_read is similar, it will
do a single decrement with the lock held") but it applies to all other
callers too.  It's tricky enough that it should be documented in the code.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Reviewed-by: Andrea Arcangeli <aarcange@redhat.com>
Cc: Chris Mason <clm@fb.com>
Cc: Davide Libenzi <davidel@xmailserver.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
(cherry picked from commit a484c3dd9426e1f450c7ed8ad5a2002ea1b309ea)
Signed-off-by: Rafael Aquini <aquini@redhat.com>