[kernel] cgroup: pids: rhel-specific hack to fix zombie accounting
Message-id: <20160324232602.GA25596@redhat.com>
Patchwork-id: 139287
O-Subject: [RHEL7 PATCH V2 BZ 1265339 7/8] cgroup: pids: rhel-specific hack to fix zombie accounting
Bugzilla: 1265339
RH-Acked-by: Rafael Aquini <aquini@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

RHBZ: 1265339

cgroup_pids does pids_uncharge() on exit and this is very wrong, a zombie
task must not escape the accounting. This was fixed upstream by the following
commits:

	c4eb720d172ff91764657dbef9db3469b54a7fb9
	cgroup: keep zombies associated with their original cgroups

	afcf6c8b75444382e0f9996157207ebae34a8848
	cgroup: add cgroup_subsys->free() method and use it to fix pids controller

The 2nd patch is trivial, but I do not think the 1st one is suitable for RHEL7.
Not only it needs the nontrivial changes in the generic cgroups code, not only
we would need to adapt these changes to RHEL7 because the code differs very much.

put_css_set() from __put_task_struct() is not safe in RHEL7 because the locking
in cgroup.c is not softirq-safe, so (for example) delayed_put_task_struct() rcu
callback can deadlock trying to take css_set_lock if it is called by RCU_SOFTIRQ.
Of course, this is fixable, but this needs more incompatible changes, or we will
have to backport a lot more (imo too many) patches to match upstream.

We could change put_css_set() to use workqueues if atomic_add_unless() fails, but
this means another rhel-speicific hack. And cgroup_scan_tasks()->put_task_struct()
doesn't look safe, this is called under cgroup_iter_start() which takes this lock
for reading.

So this patch adds the horrible but simple rhel-specific hack. pids_exit() just
pins and records css in task->cg_list.prev, the new helper cgroup_pids_release()
called by release_task() does uncharge and drops the reference.

Signed-off-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>