[vfio] Enable No-IOMMU option for RHEL
Message-id: <20160330211104.13342.62360.stgit@gimli.home>
Patchwork-id: 139557
O-Subject: [RHEL7.3 kernel PATCH 13/13] vfio: Enable No-IOMMU option for RHEL
Bugzilla: 1299662 1322577
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>

Bugzilla: 1299662
Upstream: N/A

Enable No-IOMMU config option which allows VFIO to be used by
userspace drivers such as DPDK in non-IOMMU protected environments.
The intention is not to support this but to make it available for
those who are already doing this unsafely with the UIO driver.  Using
VFIO for this promotes a consistent device model and leaves more
bread crumbs for us to track if this is used for evaluating support
cases.  Use of this mode taints the kernel and logs device access in
dmesg.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Rafael Aquini <aquini@redhat.com>