[x86] perf/x86/amd: Remove l1-dcache-stores event for AMD
Message-id: <1460038562-19641-20-git-send-email-jolsa@redhat.com>
Patchwork-id: 140351
O-Subject: [PATCH RHEL7.3 BZ1323852 019/311] perf/x86/amd: Remove l1-dcache-stores event for AMD
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 9cc2617de5b9222abb39cd02e90d57dfea99c6d7
Author: Vince Weaver <vincent.weaver@maine.edu>
Date: Wed Dec 9 11:34:45 2015 -0500

description
===========
This is a long standing bug with the l1-dcache-stores generic event on
AMD machines.  My perf_event testsuite has been complaining about this
for years and I'm finally getting around to trying to get it fixed.

The data_cache_refills:system event does not make sense for l1-dcache-stores.
Maybe this was a typo and it was meant to be for l1-dcache-store-misses?

In any case, the values returned are nowhere near correct for l1-dcache-stores
and in fact the umask values for the event have completely changed with
fam15h so it makes even less sense than ever.  So just remove it.

Signed-off-by: Rafael Aquini <aquini@redhat.com>