[x86] perf/x86/intel/uncore: Clean up hardware on exit
Message-id: <1460038562-19641-47-git-send-email-jolsa@redhat.com>
Patchwork-id: 140366
O-Subject: [PATCH RHEL7.3 BZ1323852 046/311] perf/x86/intel/uncore: Clean up hardware on exit
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit a46195f1782e94ad07e7485fb15b80db7e60e7aa
Author: Thomas Gleixner <tglx@linutronix.de>
Date: Mon Feb 22 22:19:11 2016 +0000

description
===========
When tearing down the boxes nothing undoes the hardware state which was setup
by box->init_box(). Add a box->exit_box() callback and implement it for the
uncores which have an init_box() callback.

This misses the cleanup in the error exit pathes, but I cannot be bothered to
implement it before cleaning up the rest of the driver, which makes that task
way simpler.

Signed-off-by: Rafael Aquini <aquini@redhat.com>