[x86] perf/x86/intel: Fix PEBS data source interpretation on Nehalem/Westmere
Message-id: <1460038562-19641-61-git-send-email-jolsa@redhat.com>
Patchwork-id: 140382
O-Subject: [PATCH RHEL7.3 BZ1323852 060/311] perf/x86/intel: Fix PEBS data source interpretation on Nehalem/Westmere
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit e17dc65328057c00db7e1bfea249c8771a78b30b
Author: Andi Kleen <ak@linux.intel.com>
Date: Tue Mar 1 14:25:24 2016 -0800

description
===========
Jiri reported some time ago that some entries in the PEBS data source table
in perf do not agree with the SDM. We investigated and the bits
changed for Sandy Bridge, but the SDM was not updated.

perf already implements the bits correctly for Sandy Bridge
and later. This patch patches it up for Nehalem and Westmere.

Signed-off-by: Rafael Aquini <aquini@redhat.com>