[tools] perf tests: Pass the subtest index to each test routine
Message-id: <1460038562-19641-64-git-send-email-jolsa@redhat.com>
Patchwork-id: 140395
O-Subject: [PATCH RHEL7.3 BZ1323852 063/311] perf tests: Pass the subtest index to each test routine
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 721a1f53df6aad3ea941f5fe95519d0d8e02bd65
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Nov 19 12:01:48 2015 -0300

RHEL7
=====
Omitted bpf changes.

description
===========
Some tests have sub-tests we want to run, so allow passing this.

Wang tried to avoid having to touch all tests, but then, having the
test.func in an anonymous union makes the build fail on older compilers,
like the one in RHEL6, where:

  test a = {
	.func = foo,
  };

fails.

To fix it leave the func pointer in the main structure and pass the subtest
index to all tests, end result function is the same, but we have just one
function pointer, not two, with and without the subtest index as an argument.

Signed-off-by: Rafael Aquini <aquini@redhat.com>