[tools] lib traceevent: Fix output of %llu for 64 bit values read on 32 bit machines
Message-id: <1460038562-19641-91-git-send-email-jolsa@redhat.com>
Patchwork-id: 140404
O-Subject: [PATCH RHEL7.3 BZ1323852 090/311] tools lib traceevent: Fix output of %llu for 64 bit values read on 32 bit machines
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 32abc2ede536aae52978d6c0a8944eb1df14f460
Author: Steven Rostedt <rostedt@goodmis.org>
Date: Mon Nov 16 17:25:16 2015 -0500

description
===========
When a long value is read on 32 bit machines for 64 bit output, the
parsing needs to change "%lu" into "%llu", as the value is read
natively.

Unfortunately, if "%llu" is already there, the code will add another "l"
to it and fail to parse it properly.

Signed-off-by: Rafael Aquini <aquini@redhat.com>