[tools] perf stat: Create events as disabled
Message-id: <1460038562-19641-122-git-send-email-jolsa@redhat.com>
Patchwork-id: 140434
O-Subject: [PATCH RHEL7.3 BZ1323852 121/311] perf stat: Create events as disabled
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 67ccdecd09cac818146b1e153ff901cb67570012
Author: Jiri Olsa <jolsa@kernel.org>
Date: Thu Dec 3 10:06:44 2015 +0100

description
===========
Currently we have 2 kinds of stat counters based on when the event is
enabled:

  1) tracee command events, which are enable once the
     tracee executes exec syscall (enable_on_exec bit)
  2) all other events which get alive within the
     perf_event_open syscall

And 2) case could raise a problem in case we want additional filter to
be attached for event. In this case we want the event to be enabled
after it's configured with filter.

Changing the behaviour of 2) events, so they all are created as disabled
(disabled bit). Adding extra enable call to make them alive once they
finish setup.

Signed-off-by: Rafael Aquini <aquini@redhat.com>