[tools] perf tui: Change default selection background color to yellow
Message-id: <1460038562-19641-125-git-send-email-jolsa@redhat.com>
Patchwork-id: 140462
O-Subject: [PATCH RHEL7.3 BZ1323852 124/311] perf tui: Change default selection background color to yellow
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 8d7d377c2bea16afa6b600a4517615a9eebb259b
Author: Ingo Molnar <mingo@kernel.org>
Date: Thu Mar 5 11:32:13 2015 +0100

description
===========
Boris reported that 'perf top' is unusable on his default 'black on
white' terminal, which uses (eye friendly) light-grey as a background
color.

The reason is that the TUI cursor for the current selection line uses
HE_COLORSET_SELECTED, and that has a default background color of
'lightgrey' - which is a common terminal background choice and thus
the colors conflict.

Use yellow as the background color instead: that should be an uncommon
terminal background, yet it's still ergonomic on both black and
white/grey terminals.

[ It would be a better solution to straight out detect color
  collisions and resolve them reasonably by converting them to RGB and
  calculating color space distances, but I was unable to find
  proper documentation for SLtt_get_color_object() to recover the
  current color scheme so I gave up ... Yellow works well enough. ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>