[tools] perf hists: Fix hists_evsel to release hists
Message-id: <1460038562-19641-128-git-send-email-jolsa@redhat.com>
Patchwork-id: 140469
O-Subject: [PATCH RHEL7.3 BZ1323852 127/311] perf hists: Fix hists_evsel to release hists
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 17577decb2ddae28f5a449ddb79cf0ed3e2312c5
Author: Masami Hiramatsu <masami.hiramatsu.pt@hitachi.com>
Date: Wed Dec 9 11:11:29 2015 +0900

description
===========
Since hists__init doesn't set the destructor of hists_evsel (which is an
extended evsel structure), when hists_evsel is released, the extended
part of the hists_evsel is not deleted (note that the hists_evsel object
itself is freed).

This fixes it to add a destructor for hists__evsel and to set it up.

Signed-off-by: Rafael Aquini <aquini@redhat.com>