[tools] perf top: Do show usage message when failing to create cpu/thread maps
Message-id: <1460038562-19641-140-git-send-email-jolsa@redhat.com>
Patchwork-id: 140480
O-Subject: [PATCH RHEL7.3 BZ1323852 139/311] perf top: Do show usage message when failing to create cpu/thread maps
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit f8a5c0b24b8b1e77a0812b0c8251db0afc0524b7
Author: Arnaldo Carvalho de Melo <acme@redhat.com>
Date: Thu Dec 10 14:48:45 2015 -0300

description
===========
This is necessary to get rid of the browser dependency from
usage_with_options() and its friends.  Because we validate the targets
which are used to create the cpu/thread maps and inform the user about
any override performed via the chosen UI, we don't need to call the
usage routine for that.

Signed-off-by: Rafael Aquini <aquini@redhat.com>