[tools] perf tools: Add event_update user level event
Message-id: <1460038562-19641-194-git-send-email-jolsa@redhat.com>
Patchwork-id: 140527
O-Subject: [PATCH RHEL7.3 BZ1323852 193/311] perf tools: Add event_update user level event
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit ffe777254cce24fb5fde3f0aa91fc755cfb1b812
Author: Jiri Olsa <jolsa@kernel.org>
Date: Sun Oct 25 15:51:36 2015 +0100

description
===========
It'll serve as a base event for additional event attributes details,
that are not part of the attr event.

At the moment this event is just a dummy one without any specific
functionality. The type value will distinguish the update event details.
It'll come in the following patches.

The idea for this event is to be extensible for any update that the
event might need in the future.

Signed-off-by: Rafael Aquini <aquini@redhat.com>