[tools] perf top: Create the evlist sooner
Message-id: <1460038562-19641-223-git-send-email-jolsa@redhat.com>
Patchwork-id: 140546
O-Subject: [PATCH RHEL7.3 BZ1323852 222/311] perf top: Create the evlist sooner
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 54f8f40384ab940e15585afde5c278c8e7726214
Author: Namhyung Kim <namhyung@kernel.org>
Date: Wed Dec 23 02:07:01 2015 +0900

description
===========
This is a preparation to support dynamic sort keys for tracepoint
events.  Dynamic sort keys can be created for specific fields in trace
events so it needs the event information, so we need to pass the evlist
to the sort routines, create it sooner so that the next patch can do
that.

[ Split from the patch passing the evlist to the sort routines ]

Signed-off-by: Rafael Aquini <aquini@redhat.com>