[tools] perf pmu: fix alias->snapshot missing initialization bug
Message-id: <1460038562-19641-249-git-send-email-jolsa@redhat.com>
Patchwork-id: 140579
O-Subject: [PATCH RHEL7.3 BZ1323852 248/311] perf pmu: fix alias->snapshot missing initialization bug
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 84530920de3c6ccb92c6661da784f6cdb66d3304
Author: Stephane Eranian <eranian@google.com>
Date: Wed Jan 6 19:50:01 2016 +0100

description
===========
This patch fixes a bug in __perf_pmu__new_alias() whereby the
alias->snapshot field was not initialized to false. This led to random
alias->snapshot value for an alias and was breaking some measurements
such as:

  $ perf stat -a -e uncore_imc/data_reads/ -I 1000 sleep 100

Because the event ended up being treated as snapshot mode, when it is
not.

Signed-off-by: Rafael Aquini <aquini@redhat.com>