[tools] perf report: Change default to use event group view
Message-id: <1460038562-19641-261-git-send-email-jolsa@redhat.com>
Patchwork-id: 140582
O-Subject: [PATCH RHEL7.3 BZ1323852 260/311] perf report: Change default to use event group view
Bugzilla: 1323852
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Tony Camuso <tcamuso@redhat.com>

Bugzilla: 1323852
https://bugzilla.redhat.com/show_bug.cgi?id=1323852

upstream
========
commit 1e9abf8b03c8f9352f54171647296c41317679a4
Author: Namhyung Kim <namhyung@kernel.org>
Date: Sun Nov 29 23:24:17 2015 +0900

description
===========
The event group view feature is to see related events together.  To use
the group view, events should be recorded as a group with a dedicated
syntax of surrounding events by braces (-e '{ evt1, evt2, ... }').

Also 'perf report' also requires the --group option to enable it.
However it's almost always beneficial to use the group view to see the
group events as it's more expressive.  And I think it's more natural to
see events together if they are recorded as a group.

Thus this patch changes the default value to enable it.  If users don't
want to see like it and keep the original behavior, they can set the
report.group config variable to false and/or use --no-group option in
the 'perf report' command line.

Signed-off-by: Rafael Aquini <aquini@redhat.com>