[block] nvme: move chardev and sysfs interface to common code
Message-id: <1461247929-26697-99-git-send-email-dmilburn@redhat.com>
Patchwork-id: 141676
O-Subject: [RHEL7.3 PATCH BZ 1288601 v2 098/158] nvme: move chardev and sysfs interface to common code
Bugzilla: 1288601
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Jarod Wilson <jarod@redhat.com>
RH-Acked-by: Jeremy McNicoll <jmcnicol@redhat.com>

From: Christoph Hellwig <hch@lst.de>

For this we need to add a proper controller init routine and a list of
all controllers that is in addition to the list of PCIe controllers,
which stays in pci.c.  Note that we remove the sysfs device when the
last reference to a controller is dropped now - the old code would have
kept it around longer, which doesn't make much sense.

This requires a new ->reset_ctrl operation to implement controleller
resets, and a new ->write_reg32 operation that is required to implement
subsystem resets.  We also now store caches copied of the NVMe compliance
version and the flag if a controller is attached to a subsystem or not in
the generic controller structure now.

Signed-off-by: Christoph Hellwig <hch@lst.de>
[Fixes for pr merge]
Signed-off-by: Keith Busch <keith.busch@intel.com>
Signed-off-by: Jens Axboe <axboe@fb.com>

(cherry picked from commit f3ca80fc11c3af566eacd99cf821c1a48035c63b)
Signed-off-by: Rafael Aquini <aquini@redhat.com>

Conflicts:
	drivers/nvme/host/pci.c

RHEL notes:

In RHEL7, io-64-nonatomic-lo-hi.h is not in include/linux subdir
and include idr.h in core.c

Signed-off-by: David Milburn <dmilburn@redhat.com>